Early file management work
Added placeholder header files for ChunkManager and Region. Region will
be a data structure for handling chunks in a more file-efficient format
(plan is to store 16x16 chunks on disk per region). Chunk manager will
be the World <=> Region interface so the World class doesn't have to
manage region but rather do ops on chunks through a standard protocol.